requirements.txt
Django==4.0.5
django-cors-headers==3.13.0
djangorestframework==3.13.1
environs==9.5.0
